% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpSpect.R
\docType{package}
\name{ExpSpect}
\alias{ExpSpect}
\alias{ExpSpect-package}
\title{ExpSpect: A package for gene-expression based 'spectrometry'}
\description{
The ExpSpect package facilitates deconvolution of gene expression
data to identify perturbations related to specific gene pathways
or druggable targets.  Used in the context of gene expression
profiling of cells treated with some unknown material of interest
(e.g. serum from patients with unknown environmental exposures),
such analysis is analagous to tandem mass spec proteomics--excpet
that rather than a spectrometer, gene expression is used together
with libraries of expression signatures (such as LINCS) to
deconvolute the composition of the material.
}
\details{
We are in no way affiliated with the LINCS project.  This package
does not provide LINCS data, which can only be obtained with
permission from the LINCS project.  Once you have access, downloading the data is straightforward
(and illustrated in a soon to be release vignette for this package).
While not providing the data, this package \strong{does} provide
tools that greatly facilitate working with these large data files
}

