% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LINCS.R
\docType{class}
\name{LINCS}
\alias{LINCS}
\title{LINCS class}
\format{\code{\link{LINCS}} class generator}
\usage{
\code{lincs = LINCS$new(dataFile, infoFile)}
\code{ids <- lincs$getProbes()[1:3]}
}
\arguments{
\item{dataFile}{path to the level for LINCS data file in hdf5 format (.gctx)}

\item{infoFile}{path to the instance info data file for the LINCS data file \emph{in hdf5 format (.gctx)}. \strong{Note:}
LINCS provides this data in a tab delimited format (inst.info).  However that file is unwieldy to work with
in that format as it takes a long time to load and takes a considerable chunk of memory.  It is much
more efficient to convert it to an hdf5 file with a structure parallel to that of the datafile.  This class
provides a utility function \code{info2hdf5} that will accomplish this for you.  See methods section below.}
}
\description{
A class to facilitate working with the level 4 data from the LINCS project.  This
class does not provide any of the data (which must be obtained under individual agreement with
LINCS).  However, once obtained, working with their large binary data files will is facilitated
by this class.
}
\section{Methods}{

\describe{
  \item{\code{getProbes()}}{Returns a vector of
      \href{http://www.bioconductor.org/packages/release/data/annotation/html/hgu133plus2.db.html}{HG
       U133 Plus 2.0} probesets for which data has been imputed from the 978
       landmark genes measured directly on the luminex bead arrays by the LINCS }
}
}
\keyword{data}

