% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpSpect.R
\docType{class}
\name{ExpSpect}
\alias{ExpSpect}
\title{ExpSpect class}
\format{\code{\link{ExpSpect}} class generator}
\usage{
\code{exps = ExpSpect$new()}
}
\description{
A class to perform expression based 'spectrometry'.
}
\section{Methods}{

\describe{
 \item{\code{calcScores(data, treated, untreated, sigs, method=c('ks', 'excos'), normalize=TRUE, ...)}}{Given a gene expression matrix \code{data}, computes fold change based on
      columns given in vectors \code{treated} and \code{untreated}, and then computes enrichment scores against each column of
      matrix \code{sigs}.  The \code{sigs} matrix is simply a matrix of gene expression data, one column per pertubagen which
      can be generated in any number of ways--assembled from GEO, downloaded from CMAP, extracted from LINCS, etc.  See
      Vignette for an example.  Data will be renormalized after fold change calculation prior to processing by \code{method} if \code{normalize=TRUE}.
      Extra-parameters (\code{...}) will be sent to selected function (see below for details)}
 \item{\code{cmap(data, y=NULL, up, down)}}{Given a gene expression matrix \code{data}, computes the connectivity
      score based on either a vector of expression data (\code{y}), or explicit sets of \code{up} and \code{down}
      regulated genes pertinent to a biological state or perturbation of interest.  Returns the list of connectivity
      scores for each column of \code{data}. If providing a vector of expression values, up and down genes are s
      selected based on the provided \code{threshold}, after optional normalization (\code{normalize = TRUE}) step.}
}
}
\keyword{data}

